Adds local source for fonts
Adds local source for fonts to allow for installing the fonts on a machine and still be able to access without downloading. This is helpful in situations where a security policy prohibits downloading fonts. Installing is typically used in this case as a workaround but requires a css change to be used.